pyproject.toml
[project]
name = "mscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
]
